requirements.txt
Django==1.7.1
django-shopify-auth==0.3.0
python-dotenv==0.1.0
